@@ -1,13 +1,17 @@
-module gocon
+module joyku
 
-go 1.22.0
+go 1.23.0
+
+toolchain go1.23.4
 
 require (
+	github.com/godbus/dbus/v5 v5.1.0
 	github.com/spf13/viper v1.19.0
 	github.com/sstallion/go-hid v0.14.1
 )
 
 require (
+	github.com/a-h/templ v0.3.857
 	github.com/fsnotify/fsnotify v1.7.0 // indirect
 	github.com/hashicorp/hcl v1.0.0 // indirect
 	github.com/magiconair/properties v1.8.7 // indirect
@@ -23,7 +27,7 @@ require (
 	go.uber.org/atomic v1.9.0 // indirect
 	go.uber.org/multierr v1.9.0 // indirect
 	golang.org/x/exp v0.0.0-20230905200255-921286631fa9 // indirect
-	golang.org/x/sys v0.18.0 // indirect
+	golang.org/x/sys v0.31.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
 	gopkg.in/ini.v1 v1.67.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
